+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+
+####1.0.1
+* [create] spring-security tags
+* [create] common-UUID/randomString
+
 ####1.0.0
 * [create] spring-mvc tags for i18n
 * [create] spring-mvc functions for common